crud with auth cmplt

+++ server/controllers/todoController.js
@@ -1,14 +1,14 @@
 const asyncHandler = require("express-async-handler");
 const Todo = require("../models/todoModel");
+const User = require("../models/userModel");
 // @desp GET goals
 // @route GET /api/todos
 // @access PRIVATE
 
 const getTodo = asyncHandler(async (req, res) => {
-  const tasks = await Todo.find();
+  const tasks = await Todo.find({ user: req.user.id });
   res.json(tasks);
 });
-
 
 // @desp GET goal by id
 // @route GET /api/todos
@@ -28,6 +28,7 @@ const postTodo = asyncHandler(async (req, res) => {
     throw new Error("Please add a text");
   }
   const tasks = await Todo.create({
+    user: req.user.id,
     task: req.body.task,
     desp: req.body.desp,
   });
@@ -44,6 +45,20 @@ const updateTodo = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Task not found");
   }
+
+  const user = await User.findById(req.user.id);
+  // check for user
+  if(!user){
+    res.status(401);
+    throw new Error("User not found")
+  }
+
+  // the login user matches the task user
+  if(task.user.toString() !== user.id){
+    res.status(401);
+    throw new Error("User doest not match")
+  }
+
   const updatedTodo = await Todo.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
@@ -55,13 +70,27 @@ const updateTodo = asyncHandler(async (req, res) => {
 // @access PRIVATE
 
 const deleteTodo = asyncHandler(async (req, res) => {
-  const task = await Todo.findById(req.params.id)
-  if(!task){
+  const task = await Todo.findById(req.params.id);
+  if (!task) {
     res.status(400);
     throw new Error("Task not found");
   }
-  const deletedTodo = await Todo.findByIdAndDelete(req.params.id)
-  res.json({ message: `Delete Goal ${req.params.id}` });
+  
+  const user = await User.findById(req.user.id);
+  // check for user
+  if(!user){
+    res.status(401);
+    throw new Error("User not found")
+  }
+
+  // the login user matches the task user
+  if(task.user.toString() !== user.id){
+    res.status(401);
+    throw new Error("User doest not match")
+  }
+
+  const deletedTodo = await Todo.findByIdAndDelete(req.params.id);
+  res.json(deletedTodo);
 });
 
 module.exports = {
